refactor: further cleanup to prepare for Kraken
